# packages/server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Eigen Engine API",
    "version": "1.0.0",
    "description": "The server-authoritative game engine API. Client routes (`/api/engine/*`) require a Firebase ID token; the external-bot webhook (`/api/bot/action`) is HMAC-authenticated (§7)."
  },
  "security": [
    {
      "firebase": []
    }
  ],
  "components": {
    "securitySchemes": {
      "firebase": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "A Firebase ID token, sent as `Authorization: Bearer <token>` (or `?token=` on WebSocket upgrades)."
      },
      "botHmac": {
        "type": "apiKey",
        "in": "header",
        "name": "Eigen-Signature",
        "description": "An external bot's HMAC signature over the exact request body, bound to the `action` domain (§7). Scheme `v1,<base64>`; the per-bot key is `HMAC(BOT_SIGNING_SECRET, bot_id)`. The engine signs wakes with the same header in the other direction."
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "code": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "Created": {
        "type": "object",
        "properties": {
          "game_id": {
            "type": "string"
          },
          "short_code": {
            "type": "string"
          }
        },
        "required": [
          "game_id",
          "short_code"
        ]
      },
      "CreateGame": {
        "type": "object",
        "properties": {
          "access": {
            "type": "string",
            "enum": [
              "public",
              "private",
              "friends"
            ],
            "default": "public"
          },
          "schema_version": {
            "type": "integer"
          },
          "config": {
            "type": "object",
            "additionalProperties": {},
            "default": {}
          },
          "min_players": {
            "type": "integer",
            "minimum": 1
          },
          "max_players": {
            "type": "integer",
            "minimum": 1
          },
          "rated": {
            "type": "boolean"
          },
          "turn_seconds": {
            "type": [
              "integer",
              "null"
            ],
            "exclusiveMinimum": 0,
            "default": null
          },
          "budget_seconds": {
            "type": [
              "integer",
              "null"
            ],
            "exclusiveMinimum": 0,
            "default": null
          },
          "increment_seconds": {
            "type": [
              "integer",
              "null"
            ],
            "exclusiveMinimum": 0,
            "default": null
          }
        },
        "required": [
          "schema_version",
          "min_players",
          "max_players"
        ]
      },
...
      "Join": {
        "type": "object",
        "properties": {
          "client_schema_version": {
            "type": "integer"
          },
          "command_id": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          }
        },
        "required": [
          "client_schema_version"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/api/engine/games": {
      "post": {
        "operationId": "createGame",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGame"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The created game",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Created"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "State conflict — resync and retry",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Assertion mismatch",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/engine/games/solo": {
      "post": {
        "operationId": "createSoloGame",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSolo"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The created-and-started solo game",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SoloStarted"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "State conflict — resync and retry",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Assertion mismatch",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/engine/games/{gameId}/join": {
      "post": {
        "operationId": "joinGame",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "required": true,
            "name": "gameId",
            "in": "path"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Join"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Seated — the post-join roster",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LobbyAccepted"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "State conflict — resync and retry",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Assertion mismatch",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/engine/games/join-by-code": {
      "post": {
        "operationId": "joinGameByCode",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JoinByCode"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Seated — the post-join roster",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LobbyAccepted"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "State conflict — resync and retry",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Assertion mismatch",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}